feat: add Jest Test back into setup
I now understand why Jest Test wasn't working as expected.
I did not understand why linkedList.size kept returning 0 when it
should be returning 3 (the updated size). 0 was the initialized value.
To get the updated value, I needed to create a method such as
linkedList.getSize(). I now know that this misunderstanding was not due
to a bug with Jest, but a particular behaviour of objects.